+++ src/test/kotlin/net/pwall/json/schema/codegen/CodeGeneratorNestedClassTest.kt
@@ -31,13 +31,11 @@ import kotlin.test.expect
 import java.io.File
 import java.io.StringWriter
 
-import net.pwall.log.LoggerFactory
-
 class CodeGeneratorNestedClassTest {
 
     @Test fun `should output deeply nested class`() {
         val input = File("src/test/resources/test-nested-object.schema.json")
-        val codeGenerator = CodeGenerator(loggerFactory = LoggerFactory.getDefault())
+        val codeGenerator = CodeGenerator()
         codeGenerator.baseDirectoryName = "dummy"
         val stringWriter = StringWriter()
         codeGenerator.outputResolver =
@@ -49,8 +47,7 @@ class CodeGeneratorNestedClassTest {
 
     @Test fun `should output deeply nested class in Java`() {
         val input = File("src/test/resources/test-nested-object.schema.json")
-        val codeGenerator = CodeGenerator(templates = "java", suffix = "java",
-                loggerFactory = LoggerFactory.getDefault())
+        val codeGenerator = CodeGenerator(templates = "java", suffix = "java")
         codeGenerator.baseDirectoryName = "dummy"
         val stringWriter = StringWriter()
         codeGenerator.outputResolver =